% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{atlanta}
\alias{atlanta}
\title{Atlanta Homicides 1980s-90s}
\format{A data frame with 90 rows and 23 variables
 \describe{
  \item{NAME}{county name}
  \item{STATE_NAME}{state name}
  \item{STATE_FIPS}{state FIPS code (character)}
  \item{CNTY_FIPS}{county FIPS code (character)}
  \item{FIPS}{combined state and county FIPS code (character)}
  \item{FIPSNO}{fips code as numeric variable}
  \item{HR7984}{homicide rate per 100,000 (1979-84)}
  \item{HR8488}{homicide rate per 100,000 (1984-88)}
  \item{HR8893}{homicide rate per 100,000 (1988-93)}
  \item{HC7984}{homicide count (1979-84)}
  \item{HC8488}{homicide count (1984-88)}
  \item{HC8893}{homicide count (1988-93)}
  \item{PO7984}{population total (1979-84)}
  \item{PO8488}{population total (1984-88)}
  \item{PO8893}{population total (1988-93)}
  \item{PE77}{police expenditures per capita, 1977}
  \item{PE82}{police expenditures per capita, 1982}
  \item{PE87}{police expenditures per capita, 1987}
  \item{RDAC80}{resource deprivation/affluence composite variable, 1980}
  \item{RDAC85}{resource deprivation/affluence composite variable, 1985}
  \item{RDAC90}{resource deprivation/affluence composite variable, 1990}
}}
\source{
\url{https://geodacenter.github.io/data-and-lab//atlanta_old/}
}
\usage{
data(atlanta)
}
\description{
Homicides and selected socio-economic characteristics for counties surrounding Atlanta, GA. Data aggregated for three time periods: 1979-84 (steady decline in homicides), 1984-88 (stable period), and 1988-93 (steady increase in homicides).
}
\keyword{datasets}
