% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cleveland}
\alias{cleveland}
\title{Cleveland Home Sales (2015)}
\format{A data frame with 205 rows and 9 variables
 \describe{
  \item{unique_id}{unique parcel id}
  \item{parcel}{unique parcel number}
  \item{x}{point latitude}
  \item{y}{point longitude}
  \item{sale_price}{price paid for the house ($)}
  \item{tract10int}{License plate number and sometimes a description (state, color). Some entries did not include a plate number.}
  \item{Quarter}{quarter of sale (4th for all)}
  \item{year1}{year of sale (2015 for all)}
  \item{yrquarter}{year and quarter of sale (4th quarter of 2015 for all)}
}}
\source{
\url{https://geodacenter.github.io/data-and-lab//clev_sls_154_core/}
}
\usage{
data(cleveland)
}
\description{
Location and sales price of home sales in a core area of Cleveland, OH for the fourth quarter of 2015.
}
\keyword{datasets}
