% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ncovr}
\alias{ncovr}
\title{US county homicides 1960-1990}
\format{An object of class \code{SpatialPolygonsDataFrame} with 3085 rows and 69 columns.}
\source{
\url{https://geodacenter.github.io/data-and-lab/natregimes/}
}
\usage{
data(ncovr)
}
\description{
Homicides and selected socio-economic characteristics for continental U.S. counties. Data for four decennial census years: 1960, 1970, 1980 and 1990.
}
\keyword{datasets}
