% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{nepal}
\alias{nepal}
\title{Health, poverty and education indicators for Nepal districts}
\format{An object of class \code{SpatialPolygonsDataFrame} with 75 rows and 61 columns.}
\source{
\url{https://geodacenter.github.io/data-and-lab/nepal/}
}
\usage{
data(nepal)
}
\description{
Contains development-related data for 75 districts in Nepal.
}
\keyword{datasets}
