% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{nyc_earnings}
\alias{nyc_earnings}
\title{Block-level Earnings in NYC (2002-14)}
\format{A data frame with 108487 rows and 70 variables
 \describe{
  \item{YR}{02 to 14 (2002 to 2014)}
  \item{GEOID10}{unique ID (FIPS code)}
  \item{C000_YR}{Total number of jobs}
  \item{CE01_YR}{Number of jobs with earnings $1250/month or less}
  \item{CE02_YR}{Number of jobs with earnings $1251/month to $3333/month}
  \item{CE03_YR}{Number of jobs with earnings greater than $3333/month}
}}
\source{
\url{https://geodacenter.github.io/data-and-lab//LEHD_Data/}
}
\usage{
data(nyc_earnings)
}
\description{
Block-level Earnings New York City (2002-14)
}
\keyword{datasets}
