% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{nyc_educ}
\alias{nyc_educ}
\title{NYC Education (2000)}
\format{An object of class \code{SpatialPolygonsDataFrame} with 2216 rows and 56 columns.}
\source{
\url{https://geodacenter.github.io/data-and-lab//NYC-Census-2000/}
}
\usage{
data(nyc_educ)
}
\description{
Source for school locations.
}
\keyword{datasets}
